Make Firebase storable a class to improve dry
