+++ FireRecord/Source/FirebaseFile.swift
@@ -10,12 +10,4 @@ import HandyJSON
 
 open class FirebaseFile: FirebaseStorable {
     
-    public var path: String?
-    public var data: Data?
-    
-    public required init() {}
-    
-    public func mapping(mapper: HelpingMapper) {
-        mapper >>> data
-    }
 }
